requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
anytree==2.13.0
ascii_colors==0.11.3
asgiref==3.8.1
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
autograd==1.8.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.3.0
beartype==0.18.5
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
camelot-py==1.0.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chromadb
click>=8.0.1,<8.1.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configparser==7.2.0
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
docker==4.2.0
dockerpty==0.4.1
docopt==0.6.2
docx==0.2.4
docy==0.0.0.0
dotenv==0.9.9
duckdb==1.3.0
durationpy==0.9
ecdsa==0.19.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.9
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.0
google-auth==2.40.1
googleapis-common-protos==1.70.0
graspologic==3.4.1
graspologic-native==1.2.5
greenlet==3.2.1
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.71.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
humanfriendly==10.0
hyppo==0.4.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.12.3
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jq==1.8.0
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-tavily==0.1.6
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langdetect==1.0.9
langsmith==0.3.42
lightrag-hku==1.3.7
llvmlite==0.44.0
lxml==5.4.0
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
nano-vectordb==0.0.4.3
narwhals==1.37.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnx==1.18.0
playwright
onnxruntime==1.18.0
openai==1.77.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.33.0
opentelemetry-exporter-otlp-proto-common==1.33.0
opentelemetry-exporter-otlp-proto-grpc==1.33.0
opentelemetry-instrumentation==0.54b0
opentelemetry-instrumentation-asgi==0.54b0
opentelemetry-instrumentation-fastapi==0.54b0
opentelemetry-proto==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
opentelemetry-util-http==0.54b0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
patsy==1.0.1
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pi_heif==0.22.0
pickleshare==0.7.5
pillow==11.2.1
pinecone==7.0.1
pinecone-client==3.0.2
pinecone-plugin-assistant==1.6.0
pinecone-plugin-interface==0.0.7
pipmaster==0.7.2
pipreqs==0.5.0
platformdirs==4.3.7
popper
posthog==4.0.1
POT==0.9.5
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf
protoc-gen-openapiv2==0.0.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
#pyasn1==0.4.8
#pyasn1_modules==0.4.2
#pycparser==2.22
#pydantic==2.11.4
#pydantic-settings==2.9.1
#pydantic_core==2.33.2
#pydeck==0.9.1
#Pygments==2.19.1
#PyJWT==2.10.1
#pykwalify==1.7.0
#pymssql==2.3.4
PyMuPDF==1.25.5
pynndescent==0.5.13
pyodbc==5.2.0
pyparsing==3.2.3
pypdf==5.5.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyspellchecker==0.8.3
python-box==4.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-tds==1.16.1
pytz==2025.2
pyuca==1.2
#pyvis==0.3.2
#pywin32                # vielleicht löschen
PyYAML
pyzmq==26.4.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
ruamel.yaml==0.18.11
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
semver==3.0.4
sentence-transformers==4.1.0
setuptools==80.4.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spython==0.0.79
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.4
streamlit==1.45.0
streamlit-option-menu==0.4.0
sympy==1.14.0
tabulate==0.9.0
tavily-python==0.7.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.3
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
unstructured==0.17.2
unstructured-client==0.34.0
unstructured-inference==0.8.10
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
XlsxWriter==3.2.3
xmltodict==0.14.2
yarg==0.1.9
yarl==1.20.0
zeep==4.3.1
zipp==3.21.0
zstandard==0.23.0
